+++ test/plugin/scenarios/struts2.3-scenario/pom.xml
@@ -64,6 +64,12 @@
             <version>${lombok.version}</version>
             <scope>provided</scope>
         </dependency>
+
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+            <version>4.3</version>
+        </dependency>
     </dependencies>
 
     <build>
