+++ test/plugin/scenarios/struts2.5-scenario/pom.xml
@@ -64,6 +64,11 @@
             <artifactId>struts2-core</artifactId>
             <version>${test.framework.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+            <version>4.3</version>
+        </dependency>
     </dependencies>
 
     <build>
