+++ test/plugin/scenarios/struts2.5-scenario/src/main/java/org/apache/skywalking/apm/testcase/struts/Case1Action.java
@@ -18,30 +18,11 @@
 
 package org.apache.skywalking.apm.testcase.struts;
 
-import org.apache.http.HttpEntity;
-import org.apache.http.client.ResponseHandler;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.util.EntityUtils;
-
 import java.io.IOException;
 
-public class CaseAction {
+public class Case1Action {
 
     public String execute() throws IOException {
-        visit();
         return "SUCCESS";
-    }
-
-    private static void visit() throws IOException {
-        try (CloseableHttpClient httpClient = HttpClients.createDefault()) {
-            HttpGet httpget = new HttpGet("http://localhost:8080/struts2.3-scenario/struts-scenario/case1.action");
-            ResponseHandler<String> responseHandler = response -> {
-                HttpEntity entity = response.getEntity();
-                return entity != null ? EntityUtils.toString(entity) : null;
-            };
-            httpClient.execute(httpget, responseHandler);
-        }
     }
 }
